Rename files to TreeView
